backend/requirements.txt
flask
flask-cors
python-dotenv
gunicorn
mysql-connector-python